# openapi-schema.yml
openapi: 3.0.3
info:
  title: NoCode Backend API
  version: 1.0.0
  description: API pour la plateforme NoCode - Création d'applications sans code
paths:
...
  /api/v1/studio/components/:
    get:
      operationId: studio_components_list
      description: API endpoint pour gérer les composants
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - studio
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedComponentList'
          description: ''
    post:
      operationId: studio_components_create
      description: API endpoint pour gérer les composants
      tags:
      - studio
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComponentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ComponentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ComponentRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Component'
          description: ''
...
  /api/v1/studio/projects/:
    get:
      operationId: studio_projects_list
      description: "\n        **Récupère la liste de tous les projets accessibles\
        \ par l'utilisateur connecté.**\n        \n        ## \U0001F3AF Projets retournés\
        \ :\n        \n        ### Pour un Client Individuel (sans organisation) :\n\
        \        - ✅ Tous les projets qu'il a créés personnellement\n        - ✅ Les\
        \ projets où `organization` est NULL\n        \n        ### Pour un Membre\
        \ d'Organisation :\n        - ✅ Tous ses projets personnels (créés par lui,\
        \ organization=NULL)\n        - ✅ Tous les projets de ses organisations (où\
        \ il est membre actif)\n        - ✅ Visible par tous les membres (OWNER, ADMIN,\
        \ MEMBER)\n        \n        ## \U0001F4CA Réponse :\n        - Liste de projets\
        \ avec leurs détails complets\n        - Inclut : tracking_id, name, schema_name,\
        \ organization, created_by, dates\n        \n        ## \U0001F510 Permissions\
        \ :\n        - Authentification requise (Bearer Token)\n        - Retourne\
        \ uniquement les projets accessibles par l'utilisateur\n        \n       \
        \ ## \U0001F4A1 Exemple d'utilisation :\n        ```\n        GET /api/v1/studio/projects/\n\
        \        Authorization: Bearer <votre_token>\n        ```\n        "
      summary: 📋 Lister tous les projets accessibles
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - 📁 Studio - Projets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProjectList'
          description: ''
    post:
      operationId: studio_projects_create
      description: "\n        **Crée un nouveau projet NoCode.**\n        \n     \
        \   ## \U0001F3AF Types de projets :\n        \n        ### 1️⃣ Projet Personnel\
        \ (Client Individuel) :\n        ```json\n        {\n            \"name\"\
        : \"Mon Application Mobile\"\n            // Ne pas fournir organization_id\n\
        \        }\n        ```\n        - Le projet sera lié uniquement au créateur\n\
        \        - `organization` sera NULL\n        - Seul le créateur aura accès\n\
        \        \n        ### 2️⃣ Projet d'Organisation :\n        ```json\n    \
        \    {\n            \"name\": \"ERP Entreprise\",\n            \"organization_id\"\
        : \"uuid-de-votre-organisation\"\n        }\n        ```\n        - Le projet\
        \ sera lié à l'organisation\n        - Tous les membres actifs de l'organisation\
        \ auront accès\n        - OWNER/ADMIN peuvent modifier, MEMBER peut lire\n\
        \        \n        ## ⚙️ Ce qui est créé automatiquement :\n        - ✅ Un\
        \ schéma PostgreSQL unique (`schema_name`)\n        - ✅ Une page d'accueil\
        \ par défaut\n        - ✅ Une structure de base pour l'application\n     \
        \   \n        ## \U0001F510 Permissions :\n        - Authentification requise\n\
        \        - Pour créer un projet d'organisation : être membre actif de cette\
        \ organisation\n        \n        ## \U0001F4CA Réponse :\n        - Détails\
        \ complets du projet créé\n        - Code HTTP 201 Created\n        \n   \
        \     ## ⚠️ Notes importantes :\n        - Le nom du projet doit être unique\
        \ dans votre espace\n        - Le schema_name est généré automatiquement\n\
        \        - La suppression d'un projet supprime toutes ses données\n      \
        \  "
      summary: ➕ Créer un nouveau projet
      tags:
      - 📁 Studio - Projets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
            examples:
              ProjetPersonnel:
                value:
                  name: Mon Portfolio
                summary: Projet Personnel
              ProjetOrganisation:
                value:
                  name: CRM Enterprise
                  organization_id: 550e8400-e29b-41d4-a716-446655440000
                summary: Projet Organisation
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
  /api/v1/studio/projects/{id}/:
    get:
      operationId: studio_projects_retrieve
      description: "\n        **Récupère les informations détaillées d'un projet spécifique.**\n\
        \        \n        ## \U0001F3AF Informations retournées :\n        - tracking_id\
        \ (UUID unique du projet)\n        - name (nom du projet)\n        - schema_name\
        \ (nom du schéma PostgreSQL)\n        - organization (organisation liée, ou\
        \ NULL si projet personnel)\n        - created_by (email et nom du créateur)\n\
        \        - created_at / updated_at (dates de création/modification)\n    \
        \    \n        ## \U0001F510 Permissions :\n        - Authentification requise\n\
        \        - Accès autorisé si :\n          - ✅ Vous êtes le créateur du projet\n\
        \          - ✅ Vous êtes membre actif de l'organisation du projet\n      \
        \  \n        ## \U0001F4A1 Exemple :\n        ```\n        GET /api/v1/studio/projects/{tracking_id}/\n\
        \        Authorization: Bearer <votre_token>\n        ```\n        "
      summary: 🔍 Récupérer les détails d'un projet
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - 📁 Studio - Projets
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    put:
      operationId: studio_projects_update
      description: "\n        **Modifie complètement un projet existant.**\n     \
        \   \n        ## \U0001F4DD Champs modifiables :\n        - `name` : Nouveau\
        \ nom du projet\n        - `organization_id` : Changer d'organisation (si\
        \ autorisé)\n        \n        ## \U0001F510 Permissions :\n        - Authentification\
        \ requise\n        - Modification autorisée si :\n          - ✅ Vous êtes\
        \ le créateur du projet\n          - ✅ Vous êtes OWNER ou ADMIN de l'organisation\
        \ du projet\n          - ❌ Les MEMBER ne peuvent PAS modifier\n        \n\
        \        ## ⚠️ Notes :\n        - Le schema_name ne peut pas être modifié\n\
        \        - Le tracking_id ne change jamais\n        \n        ## \U0001F4A1\
        \ Exemple :\n        ```json\n        PUT /api/v1/studio/projects/{tracking_id}/\n\
        \        {\n            \"name\": \"Nouveau Nom du Projet\"\n        }\n \
        \       ```\n        "
      summary: ✏️ Modifier un projet (PUT complet)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - 📁 Studio - Projets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    patch:
      operationId: studio_projects_partial_update
      description: "\n        **Modifie partiellement un projet existant.**\n    \
        \    \n        Même fonctionnement que PUT, mais vous pouvez ne modifier que\
        \ les champs souhaités.\n        \n        ## \U0001F4A1 Exemple :\n     \
        \   ```json\n        PATCH /api/v1/studio/projects/{tracking_id}/\n      \
        \  {\n            \"name\": \"Nouveau Nom\"\n        }\n        ```\n    \
        \    \n        ## \U0001F510 Permissions : Identiques à PUT\n        "
      summary: ✏️ Modifier partiellement un projet (PATCH)
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - 📁 Studio - Projets
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProjectRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProjectRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProjectRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
    delete:
      operationId: studio_projects_destroy
      description: "\n        **Supprime définitivement un projet et toutes ses données\
        \ associées.**\n        \n        ## ⚠️ ATTENTION - Cette action est IRRÉVERSIBLE\
        \ !\n        \n        ### Ce qui sera supprimé :\n        - \U0001F5D1️ Le\
        \ projet lui-même\n        - \U0001F5D1️ Le schéma PostgreSQL complet\n  \
        \      - \U0001F5D1️ Toutes les tables du projet\n        - \U0001F5D1️ Toutes\
        \ les pages et composants\n        - \U0001F5D1️ Toutes les données stockées\
        \ dans ce projet\n        \n        ## \U0001F510 Permissions :\n        -\
        \ Authentification requise\n        - Suppression autorisée si :\n       \
        \   - ✅ Vous êtes le créateur du projet\n          - ✅ Vous êtes OWNER ou\
        \ ADMIN de l'organisation du projet\n          - ❌ Les MEMBER ne peuvent PAS\
        \ supprimer\n        \n        ## \U0001F4CA Réponse :\n        - Code HTTP\
        \ 204 No Content (succès)\n        - Code HTTP 403 Forbidden (pas autorisé)\n\
        \        - Code HTTP 404 Not Found (projet inexistant)\n        \n       \
        \ ## \U0001F4A1 Conseil :\n        Avant de supprimer, assurez-vous d'avoir\
        \ exporté les données importantes !\n        "
      summary: 🗑️ Supprimer un projet
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - 📁 Studio - Projets
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/studio/projects/{id}/schemas/:
    get:
      operationId: studio_projects_schemas_retrieve
      description: Liste des schémas de données du projet
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - studio
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
...
components:
  schemas:
...
    Component:
      type: object
      description: Alias pour ComponentInstanceSerializer pour compatibilité
      properties:
        id:
          type: integer
          readOnly: true
        page:
          type: integer
        component:
          type: integer
        parent:
          type: integer
          nullable: true
          description: Composant parent (pour layout imbriqué)
        slot:
          type: string
          description: 'Slot/zone dans le parent (ex: header, content, left, right)'
          maxLength: 100
        position:
          type: object
          additionalProperties: {}
          description: 'Position/layout libre: {x,y,w,h,breakpoints,...}'
        order:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        config:
          type: object
          additionalProperties: {}
        linked_field_schema:
          type: integer
          nullable: true
        is_auto_generated:
          type: boolean
        needs_sync:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - component
      - page
    ComponentRequest:
      type: object
      description: Alias pour ComponentInstanceSerializer pour compatibilité
      properties:
        page:
          type: integer
        component:
          type: integer
        parent:
          type: integer
          nullable: true
          description: Composant parent (pour layout imbriqué)
        slot:
          type: string
          description: 'Slot/zone dans le parent (ex: header, content, left, right)'
          maxLength: 100
        position:
          type: object
          additionalProperties: {}
          description: 'Position/layout libre: {x,y,w,h,breakpoints,...}'
        order:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        config:
          type: object
          additionalProperties: {}
        linked_field_schema:
          type: integer
          nullable: true
        is_auto_generated:
          type: boolean
        needs_sync:
          type: boolean
      required:
      - component
      - page
...
    PaginatedComponentList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Component'
...
    PaginatedProjectList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Project'
...
    PatchedProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        organization_id:
          type: string
          format: uuid
          writeOnly: true
          nullable: true
...
    Project:
      type: object
      properties:
        tracking_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        organization_name:
          type: string
          readOnly: true
        schema_name:
          type: string
          readOnly: true
        created_by_email:
          type: string
          readOnly: true
        created_by_name:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - name
    ProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        organization_id:
          type: string
          format: uuid
          writeOnly: true
          nullable: true
      required:
      - name
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
- url: http://localhost:8000
  description: Development server
- url: https://api.nocode.com
  description: Production server